Use a template for rendering storage items
We use the WinJS.Binding.Template method for templating the storage items.
This is a little tricky, since most information about the items (display
name, but also thumbnail, etc) are only available asynchronously, so the
we still need a bit of code for binding converters which register event
handlers that listen to updates to the items' properties.

The bindThumbnail() function comes directly from a VS sample
("StorageDataSource and GetVirtualizedFilesVector sample"), the
bindStorageItemProperty() function is homegrown.