Stop listening to further update to properties after first good update
When e.g. a file gets renamed, the ListView and StorageDataSource refresh
the entire item, as opposed to triggering an update of its properties
(the same goes for thumbnails when e.g. modifying the contents of a
picture), so we don't need to stay hooked to the item, we can let it
forget us, thus freeing up some resources.