Very crude attempt at nicer formatting
This is really just a cheap hack, especially for aligning columns: I have
the feeling that the ListView is not going to help us here, and that
we'll have to find something else to replace the tradition detail list
view.

Using Microsoft's implementation of flexbox. This seems to be on the way
to become a standard, so I'm fine with that.